refactor: move append-only duckdb-index helpers into the append tool
duckdb_index.py keeps only shared/init-side helpers (TablePair, read_pops_legend, read_and_validate_pops_legends, compute_joint_legend, write_metadata_tables); the sequences-row-building helpers + their tests now live with the append tool.
